why do these suites fail in parallel?
