+++ packages/integrations/image/test/picture-ssg.test.js
@@ -2,6 +2,7 @@ import { expect } from 'chai';
 import * as cheerio from 'cheerio';
 import fs from 'fs';
 import sizeOf from 'image-size';
+import slash from 'slash';
 import { fileURLToPath } from 'url';
 import { loadFixture } from './test-utils.js';
 
@@ -36,7 +37,7 @@ describe('SSG pictures - dev', function () {
 			const src = image.attr('src');
 			const [route, params] = src.split('?');
 
-			expect(route).to.equal('/@astroimage/assets/social.jpg');
+			expect(slash(route)).to.equal('/@astroimage/assets/social.jpg');
 
 			const searchParams = new URLSearchParams(params);
 
@@ -62,7 +63,7 @@ describe('SSG pictures - dev', function () {
 			const src = image.attr('src');
 			const [route, params] = src.split('?');
 
-			expect(route).to.equal('/@astroimage/assets/social.jpg');
+			expect(slash(route)).to.equal('/@astroimage/assets/social.jpg');
 
 			const searchParams = new URLSearchParams(params);
